Module stn::stn

Structs

Edge

RegistrationEnum
(node count, edge count) tuple struct

RegistrationOptions

RegistrationPayload

STN

Constants

_IMPL_DESERIALIZE_FOR_Edge

_IMPL_DESERIALIZE_FOR_RegistrationOptions

_IMPL_DESERIALIZE_FOR_RegistrationPayload

_IMPL_SERIALIZE_FOR_Edge

_IMPL_SERIALIZE_FOR_RegistrationOptions

_IMPL_SERIALIZE_FOR_RegistrationPayload

Functions

build_distance_graph
Build the distance graph. Returns (#nodes, #edges) tuple

commit_and_tighten
Commit an as-performed time to the STN. Updates bounds on remaining activities. Errs if the as_performed time creates a conflict

default_execution_uncertainty
Default uncertainty for the edge between two nodes if the interval is not given

initialize
Initialize the STN using APSP set activity bounds

perform_apsp
Perform All Pairs Shortest Paths algorithm and generate the constraint table

set_bounds
Use the constraint table to set the bounds on each node